+++ src/main/java/com/github/games647/lagmonitor/traffic/CleanUpTask.java
@@ -5,12 +5,6 @@ import io.netty.channel.ChannelPipeline;
 
 import java.util.NoSuchElementException;
 
-/**
- * Moving this runnable to outer class fixes class loading issues on reload of the server which ends up in a
- * ClassNotFoundException on plugin disable:
- *
- * ClassNotFoundException: com.github.games647.lagmonitor.traffic.TinyProtocol$3
- */
 public class CleanUpTask implements Runnable {
 
     private final ChannelPipeline pipeline;